added a reusable MainMenu component.
It is designed to wrap its children (e.g. this.props.children) in
a bootstrap dropdown li.  So, you can pass a context specific dropdown
menu to it.  See how it is used in ProjectDashboard for an example.

Also added a NewProjectButton. This will eventually handle the click
of the anchor to produce a new Project Modal.  It is reused in both
the Main ProjectDashboard menu, and the MainMenu that is created for
ProjectDashboard.