modified the handleAjaxError to switch between expiredApiToken 401 status errors and 400 validation errors
added handleValidationErrors to provide a generic system to use bootstrap error classes
to style a form when errors occur. Forms must implement a {field}Field div which contains
the labels and inputs, and can be updated with the bootstrap has_error class, and a {field}Alert
div which will display an alert to the user with more information.