rake task for deploying secrets from vault to heroku
Uses the information provided by application_info.yml
  id: app_id
  target: the name of the heroku app to which the config will be deployed

Requires that the developers Host Machine ~/.netrc is mounted as a volume to /root/.netrc
 which is configured by using the heroku commandline client to login
 (see https://devcenter.heroku.com/articles/heroku-command)

added a stub ua_test section to database.yml for this to run, but it does not use the database

Requires that the secrets be published to vault using the mount /duke_data_service/#{Rails.env}
with a read-only policy be granted to an app_id using the application_info["id"] as the app_id and the
heroku application id obtained using the heroku HTTP API for the application_info["target"]

secrets should be in the form X=Y on the commandline, so that they are consumed as a JSON hash
using the Vault API and app_id authentication

dc-dev.utils.yml deploysecrets task requires that a RAILS_ENV environment variable be set on the host machine
before running it. It will not work to run the deploysecrets task with RAILS_ENV=development before or after the
command, it must be a seperate environment variable. This may be fixed in future versions