The /search endpoint is configured to behave according to the Apiary specification. It takes an Array of included_kinds and a search_query Hash. Unfortunately, the swaggerui frontend is not interpreting the information provided by grape_swagger to present these variables in the form that the user can submit to the endpoint from the swaggerui frontend.
This changes the type of the included_kinds to Array[String]. The swaggerui interprets this to provide a generic
text box to input JSON, and allows the full JSON payload for the endpoint to be sent correctly. It even presents
information about the expected JSON, but this only includes what is expected in the included_kinds. It does not
present anything to the user instructing them how to send the Elasticsearch Query. It may not be possible to make
this behave better.