refactored 'invoke_task' to be a method that takes expected_stderr and expected_stdout (default //), and expects { subject.invoke } to output these to their respective ouputs
this allows us to remove the silence_stream wrappers from our rake task specs, and the
Rails 5 specific code to use it. silence_stream is not very well documented, and may not
last past Rails 5.

NOTE: there is a bug in the graphdb_rake_spec that needs to be fixed