Warning! Major Refactor. Moved all of the SearchDocumentSerializers into a seperate Search:: namespace. This allows us to follow general conventions for naming serializers for both API serialization (not in the Search:: namespace), and Elasticsearch serialization (in the Search:: namespace).
If an object should be serialized for direct usage, e.g. returned by the API, or seralized to a search index in Elasticsearch, then its name
should reflect that with the object name and Serializer. FolderSerializer is meant to serialize a Folder to return in the API, while Search::FolderSerializer
is meant to be serialized to the folders indices in Elasticsearch.

If an object should be serialized inside of another object, especially when the object is meant to be 'summarized' by removing attributes, associations, etc.
that would normally be returned for the full object, it should be serialized with the object name and SummarySerializer. FolderSummarySerializer is meant to
add a summary of a Folder to the serialization of another object returned by the API, while Search::FolderSummarySerializer is meant to add a summary of
a Folder to the Elasticsearch index for another objects document.